improv: clean client listeners and hooks
