--- src/main/java/com/aetherteam/aetherii/client/event/hooks/PlayerRespawnClientHooks.java
@@ -1,4 +1,0 @@
-package com.aetherteam.aetherii.client.event.hooks;
-
-public class PlayerRespawnClientHooks {
-}
